reward_model/pyproject.toml
[tool.poetry]
name = "reward-model"
version = "0.1.0"
description = ""
authors = ["Andrew Green <agreen@ebi.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
polars = "^0.18.9"
catboost = "^1.2"
sentence-transformers = "^2.2.2"
pyarrow = "^12.0.1"
vega = "^4.0.0"
torch = "^2.0.1"
datasets = "^2.14.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
